refactor: extract document lifecycle
AI-Generated: ChatGPT (OpenAI)
